refactor: Unify _to_text() function into utils.py
- Add _to_text() to utils.py with comprehensive docstring
- Import _to_text from utils.py in kernel.py and fuji.py
- Remove duplicate _to_text() definitions from kernel.py and fuji.py
- The unified version supports: query, title, text, description, prompt keys
- Includes empty string check for dict values (from fuji.py version)

This reduces code duplication and ensures consistent behavior across modules.
